refactor(util): make Xoshiro256PlusPlus a self-contained utility
Follow-up to sharing the SVM's xoshiro generator with the noncomp sampler.

Move seed_from_entropy() out of svm/svm_state.cc into a new
util/xoshiro.cc so the generator no longer has an out-of-line member
defined in an SVM translation unit. The implementation has no SVM
coupling, and <random> is no longer needed in svm_state.cc.

Add a canonical next_double() to the generator and have both the SVM's
random_double() and the sampler draw through it, so the [0, 1) extraction
has one definition instead of two copies of the same bit math. The header
inline keeps the SVM hot path unchanged. No behavior change; Debug and
Release suites green.
